Start revealing the function name when the type is fully revealed
Types like `Floating a => a -> a` give us little information about
the hidden function. Those functions tend to have the same lengths
so counting the chars don't help either. Trying to guess those
functions is frankly annoying.

To solve this, as soon as the type is fully revealed, chars from
the function name are revealed in the same way (chars occurring in
both the guess and the anwser are revealed). Note that previous
guesses are not taken into account, only the guesses made when
the type was fully revealed are used.