Wait one guess before starting to reveal name letters
Before this change, if a guess resulted in revealing the full type,
it would also reveal common letters at the same time.
Now, the letters will be revealed starting from the next guess,
giving people a chance to try a guess with the full type and an
fully unrevealed name.